refactor: separação do menu em arquivos e funções
