+++ exercicios/aula_10/caixa_eletronico/app.py
@@ -1,10 +1,10 @@
-from caixa_eletronico import menus
+from caixa_eletronico.menus.principal import menu_principal
 from caixa_eletronico.bd import base
 
 
 def main():
     base.criar_banco()
-    menus.menu_principal()
+    menu_principal()
 
 
 if __name__ == '__main__':
